AG implantado.

+++ VisitaCidades/Model/Solucao.cs
@@ -4,7 +4,7 @@ namespace VisitaCidades.Model
 {
     public class Solucao
     {
-        public int Custo { get; set; }
+        public float Custo { get; set; }
 
         public IEnumerable<Rota> Rotas { get; set; }
     }
